remove now unused controllers and sate codeclimate test coverage
