+++ app/controllers/application_controller.rb
@@ -6,18 +6,22 @@ class ApplicationController < ActionController::Base
   include Pundit
   self.responder = ApplicationResponder
   respond_to :html, :json
-
   # By default authenticated users can't access controller actions.
   #
   # This is deny-by-default, and setting it here will also cover the Devise
   # controllers which will prevent logged-in users from doing stuff like
-  # requesting password resets or logging in as an Admin while already a User
+  # requesting password resets when already logged in
   #
-  # See User::BaseController and Admin::BaseController where the appropriate
-  # user models are explicitly allowed.
-  include PublicOnly
+  # Overridden in User::BaseController
+  before_action :redirect_authenticated_users
 
   protect_from_forgery with: :exception
 
-  devise_group :account, contains: %i[user]
+  def redirect_authenticated_users
+    return unless signed_in?
+
+    flash.clear
+    flash[:alert] = 'Already logged in'
+    redirect_to(root_path)
+  end
 end
