--- app/controllers/home_controller.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-class HomeController < BaseController
-  def index; end
-end
